[update] 마이페이지 커뮤니티 리팩토링
